refactor: storybook docs를 위해 타입을 한 파일에 작성
- Flex.type.ts 제거 > Flex에 props type 선언
- props 별 설명 추가
